+++ pkg/controller/logcollector/logcollector_controller.go
@@ -47,6 +47,7 @@ import (
 	relasticsearch "github.com/tigera/operator/pkg/render/common/elasticsearch"
 	rmeta "github.com/tigera/operator/pkg/render/common/meta"
 	"github.com/tigera/operator/pkg/render/common/networkpolicy"
+	rlogcollector "github.com/tigera/operator/pkg/render/logcollector"
 	"github.com/tigera/operator/pkg/render/monitor"
 	"github.com/tigera/operator/pkg/tls/certificatemanagement"
 	"github.com/tigera/operator/pkg/url"
@@ -79,7 +80,7 @@ func Add(mgr manager.Manager, opts options.ControllerOptions) error {
 	go utils.WaitToAddLicenseKeyWatch(c, opts.K8sClientset, log, licenseAPIReady)
 	go utils.WaitToAddTierWatch(networkpolicy.CalicoTierName, c, opts.K8sClientset, log, tierWatchReady)
 	go utils.WaitToAddNetworkPolicyWatches(c, opts.K8sClientset, log, []types.NamespacedName{
-		{Name: render.FluentBitPolicyName, Namespace: render.LogCollectorNamespace},
+		{Name: rlogcollector.FluentBitPolicyName, Namespace: render.LogCollectorNamespace},
 	})
 
 	if opts.MultiTenant {
@@ -129,17 +130,17 @@ func add(mgr manager.Manager, c ctrlruntime.Controller) error {
 	}
 
 	for _, secretName := range []string{
-		render.ElasticsearchEksLogForwarderUserSecret,
-		render.S3FluentBitSecretName, render.EksLogForwarderSecret,
-		render.SplunkFluentBitTokenSecretName, monitor.PrometheusClientTLSSecretName,
-		render.FluentBitTLSSecretName, render.TigeraLinseedSecret, render.VoltronLinseedPublicCert, render.EKSLogForwarderTLSSecretName,
+		rlogcollector.ElasticsearchEksLogForwarderUserSecret,
+		rlogcollector.S3FluentBitSecretName, rlogcollector.EksLogForwarderSecret,
+		rlogcollector.SplunkFluentBitTokenSecretName, monitor.PrometheusClientTLSSecretName,
+		rlogcollector.FluentBitTLSSecretName, render.TigeraLinseedSecret, render.VoltronLinseedPublicCert, rlogcollector.EKSLogForwarderTLSSecretName,
 	} {
 		if err = utils.AddSecretsWatch(c, secretName, common.OperatorNamespace()); err != nil {
 			return fmt.Errorf("log-collector-controller failed to watch the Secret resource(%s): %v", secretName, err)
 		}
 	}
 
-	for _, configMapName := range []string{render.FluentBitFilterConfigMapName, relasticsearch.ClusterConfigConfigMapName} {
+	for _, configMapName := range []string{rlogcollector.FluentBitFilterConfigMapName, relasticsearch.ClusterConfigConfigMapName} {
 		if err = utils.AddConfigMapWatch(c, configMapName, common.OperatorNamespace(), &handler.EnqueueRequestForObject{}); err != nil {
 			return fmt.Errorf("logcollector-controller failed to watch ConfigMap %s: %v", configMapName, err)
 		}
@@ -371,7 +372,7 @@ func (r *ReconcileLogCollector) Reconcile(ctx context.Context, request reconcile
 
 	// fluentBitKeyPair is the key pair fluent-bit presents to identify itself
 	httpInputServiceNames := dns.GetServiceDNSNames(render.FluentBitInputService, render.LogCollectorNamespace, r.opts.ClusterDomain)
-	fluentBitKeyPair, err := certificateManager.GetOrCreateKeyPair(r.client, render.FluentBitTLSSecretName, common.OperatorNamespace(), append([]string{render.FluentBitTLSSecretName}, httpInputServiceNames...))
+	fluentBitKeyPair, err := certificateManager.GetOrCreateKeyPair(r.client, rlogcollector.FluentBitTLSSecretName, common.OperatorNamespace(), append([]string{rlogcollector.FluentBitTLSSecretName}, httpInputServiceNames...))
 	if err != nil {
 		r.status.SetDegraded(operatorv1.ResourceCreateError, "Error creating TLS certificate", err, reqLogger)
 		return reconcile.Result{}, err
@@ -457,7 +458,7 @@ func (r *ReconcileLogCollector) Reconcile(ctx context.Context, request reconcile
 		return reconcile.Result{}, err
 	}
 
-	var s3Credential *render.S3Credential
+	var s3Credential *rlogcollector.S3Credential
 	if instance.Spec.AdditionalStores != nil {
 		if instance.Spec.AdditionalStores.S3 != nil {
 			s3Credential, err = getS3Credential(r.client)
@@ -472,7 +473,7 @@ func (r *ReconcileLogCollector) Reconcile(ctx context.Context, request reconcile
 		}
 	}
 
-	var splunkCredential *render.SplunkCredential
+	var splunkCredential *rlogcollector.SplunkCredential
 	if instance.Spec.AdditionalStores != nil {
 		if instance.Spec.AdditionalStores.Splunk != nil {
 			splunkCredential, err = getSplunkCredential(r.client)
@@ -538,16 +539,16 @@ func (r *ReconcileLogCollector) Reconcile(ctx context.Context, request reconcile
 	for _, key := range filters.InvalidKeys() {
 		invalidFilters[key] = true
 	}
-	for _, key := range []string{render.FluentBitFilterFlowName, render.FluentBitFilterDNSName} {
+	for _, key := range []string{rlogcollector.FluentBitFilterFlowName, rlogcollector.FluentBitFilterDNSName} {
 		warningKey := "fluent-bit-filter-" + key
 		if invalidFilters[key] {
-			r.status.SetWarning(warningKey, fmt.Sprintf("Ignoring the %q entry in the %s ConfigMap: it is not valid fluent-bit YAML. Filters previously written in fluentd syntax must be rewritten as a fluent-bit YAML filter list.", key, render.FluentBitFilterConfigMapName))
+			r.status.SetWarning(warningKey, fmt.Sprintf("Ignoring the %q entry in the %s ConfigMap: it is not valid fluent-bit YAML. Filters previously written in fluentd syntax must be rewritten as a fluent-bit YAML filter list.", key, rlogcollector.FluentBitFilterConfigMapName))
 		} else {
 			r.status.ClearWarning(warningKey)
 		}
 	}
 
-	var eksConfig *render.EksCloudwatchLogConfig
+	var eksConfig *rlogcollector.EksCloudwatchLogConfig
 	var eksLogForwarderKeyPair certificatemanagement.KeyPairInterface
 	if installationSpec.KubernetesProvider.IsEKS() {
 		log.Info("Managed kubernetes EKS found, getting necessary credentials and config")
@@ -564,7 +565,7 @@ func (r *ReconcileLogCollector) Reconcile(ctx context.Context, request reconcile
 				}
 
 				// eksLogForwarderKeyPair is the key pair eks-log-forwarder presents to identify itself
-				eksLogForwarderKeyPair, err = certificateManager.GetOrCreateKeyPair(r.client, render.EKSLogForwarderTLSSecretName, common.OperatorNamespace(), []string{render.EKSLogForwarderTLSSecretName})
+				eksLogForwarderKeyPair, err = certificateManager.GetOrCreateKeyPair(r.client, rlogcollector.EKSLogForwarderTLSSecretName, common.OperatorNamespace(), []string{rlogcollector.EKSLogForwarderTLSSecretName})
 				if err != nil {
 					r.status.SetDegraded(operatorv1.ResourceCreateError, "Error creating eks log forwarder TLS certificate", err, reqLogger)
 					return reconcile.Result{}, err
@@ -595,7 +596,7 @@ func (r *ReconcileLogCollector) Reconcile(ctx context.Context, request reconcile
 	// Create a component handler to manage the rendered component.
 	handler := utils.NewComponentHandler(log, r.client, r.scheme, instance)
 
-	fluentBitCfg := &render.FluentBitConfiguration{
+	fluentBitCfg := &rlogcollector.FluentBitConfiguration{
 		LogCollector:           instance,
 		S3Credential:           s3Credential,
 		SplkCredential:         splunkCredential,
@@ -617,7 +618,7 @@ func (r *ReconcileLogCollector) Reconcile(ctx context.Context, request reconcile
 		LicenseExpired:         licenseExpired,
 	}
 	// Render the fluent-bit component for Linux
-	comp := render.FluentBit(fluentBitCfg)
+	comp := rlogcollector.FluentBit(fluentBitCfg)
 
 	certificateComponent := rcertificatemanagement.Config{
 		Namespace:       render.LogCollectorNamespace,
@@ -674,7 +675,7 @@ func (r *ReconcileLogCollector) Reconcile(ctx context.Context, request reconcile
 	}
 
 	if hasWindowsNodes {
-		fluentBitCfg = &render.FluentBitConfiguration{
+		fluentBitCfg = &rlogcollector.FluentBitConfiguration{
 			LogCollector:           instance,
 			S3Credential:           s3Credential,
 			SplkCredential:         splunkCredential,
@@ -699,7 +700,7 @@ func (r *ReconcileLogCollector) Reconcile(ctx context.Context, request reconcile
 			// still Linux-only (gated on OSType), so this only affects the policy.
 			NonClusterHost: nonclusterhost,
 		}
-		comp = render.FluentBit(fluentBitCfg)
+		comp = rlogcollector.FluentBit(fluentBitCfg)
 
 		if err = imageset.ApplyImageSet(ctx, r.client, variant, comp); err != nil {
 			r.status.SetDegraded(operatorv1.ResourceUpdateError, "Error with images from ImageSet", err, reqLogger)
@@ -723,8 +724,8 @@ func (r *ReconcileLogCollector) Reconcile(ctx context.Context, request reconcile
 
 	// Check BYO certificate expiry warnings.
 	certificatemanagement.CheckKeyPairWarnings(map[string]certificatemanagement.KeyPairInterface{
-		render.FluentBitTLSSecretName:       fluentBitKeyPair,
-		render.EKSLogForwarderTLSSecretName: eksLogForwarderKeyPair,
+		rlogcollector.FluentBitTLSSecretName:       fluentBitKeyPair,
+		rlogcollector.EKSLogForwarderTLSSecretName: eksLogForwarderKeyPair,
 	}, r.status)
 
 	// Clear the degraded bit if we've reached this far.
@@ -744,81 +745,81 @@ func (r *ReconcileLogCollector) Reconcile(ctx context.Context, request reconcile
 	return reconcile.Result{RequeueAfter: graceRequeueAfter}, nil
 }
 
-func getS3Credential(client client.Client) (*render.S3Credential, error) {
+func getS3Credential(client client.Client) (*rlogcollector.S3Credential, error) {
 	secret := &corev1.Secret{}
 	secretNamespacedName := types.NamespacedName{
-		Name:      render.S3FluentBitSecretName,
+		Name:      rlogcollector.S3FluentBitSecretName,
 		Namespace: common.OperatorNamespace(),
 	}
 	if err := client.Get(context.Background(), secretNamespacedName, secret); err != nil {
 		if errors.IsNotFound(err) {
 			return nil, nil
 		}
-		return nil, fmt.Errorf("failed to read secret %q: %s", render.S3FluentBitSecretName, err)
+		return nil, fmt.Errorf("failed to read secret %q: %s", rlogcollector.S3FluentBitSecretName, err)
 	}
 
 	var ok bool
 	var kId []byte
-	if kId, ok = secret.Data[render.S3KeyIdName]; !ok || len(kId) == 0 {
+	if kId, ok = secret.Data[rlogcollector.S3KeyIdName]; !ok || len(kId) == 0 {
 		return nil, fmt.Errorf("expected secret %q to have a field named %q",
-			render.S3FluentBitSecretName, render.S3KeyIdName)
+			rlogcollector.S3FluentBitSecretName, rlogcollector.S3KeyIdName)
 	}
 	var kSecret []byte
-	if kSecret, ok = secret.Data[render.S3KeySecretName]; !ok || len(kSecret) == 0 {
+	if kSecret, ok = secret.Data[rlogcollector.S3KeySecretName]; !ok || len(kSecret) == 0 {
 		return nil, fmt.Errorf("expected secret %q to have a field named %q",
-			render.S3FluentBitSecretName, render.S3KeySecretName)
+			rlogcollector.S3FluentBitSecretName, rlogcollector.S3KeySecretName)
 	}
 
-	return &render.S3Credential{
+	return &rlogcollector.S3Credential{
 		KeyId:     kId,
 		KeySecret: kSecret,
 	}, nil
 }
 
-func getSplunkCredential(client client.Client) (*render.SplunkCredential, error) {
+func getSplunkCredential(client client.Client) (*rlogcollector.SplunkCredential, error) {
 	tokenSecret := &corev1.Secret{}
 	tokenNamespacedName := types.NamespacedName{
-		Name:      render.SplunkFluentBitTokenSecretName,
+		Name:      rlogcollector.SplunkFluentBitTokenSecretName,
 		Namespace: common.OperatorNamespace(),
 	}
 	if err := client.Get(context.Background(), tokenNamespacedName, tokenSecret); err != nil {
 		if errors.IsNotFound(err) {
 			return nil, nil
 		}
-		return nil, fmt.Errorf("failed to read secret %q: %s", render.SplunkFluentBitTokenSecretName, err)
+		return nil, fmt.Errorf("failed to read secret %q: %s", rlogcollector.SplunkFluentBitTokenSecretName, err)
 	}
 
-	token, ok := tokenSecret.Data[render.SplunkFluentBitSecretTokenKey]
+	token, ok := tokenSecret.Data[rlogcollector.SplunkFluentBitSecretTokenKey]
 	if !ok || len(token) == 0 {
 		return nil, fmt.Errorf("expected secret %q to have a field named %q",
-			render.SplunkFluentBitTokenSecretName, render.SplunkFluentBitSecretTokenKey)
+			rlogcollector.SplunkFluentBitTokenSecretName, rlogcollector.SplunkFluentBitSecretTokenKey)
 	}
 
-	return &render.SplunkCredential{
+	return &rlogcollector.SplunkCredential{
 		Token: token,
 	}, nil
 }
 
-func getFluentBitFilters(client client.Client) (*render.FluentBitFilters, error) {
+func getFluentBitFilters(client client.Client) (*rlogcollector.FluentBitFilters, error) {
 	cm := &corev1.ConfigMap{}
 	cmNamespacedName := types.NamespacedName{
-		Name:      render.FluentBitFilterConfigMapName,
+		Name:      rlogcollector.FluentBitFilterConfigMapName,
 		Namespace: common.OperatorNamespace(),
 	}
 	if err := client.Get(context.Background(), cmNamespacedName, cm); err != nil {
 		if errors.IsNotFound(err) {
 			return nil, nil
 		}
-		return nil, fmt.Errorf("failed to read ConfigMap %q: %s", render.FluentBitFilterConfigMapName, err)
+		return nil, fmt.Errorf("failed to read ConfigMap %q: %s", rlogcollector.FluentBitFilterConfigMapName, err)
 	}
 
-	return &render.FluentBitFilters{
-		Flow: cm.Data[render.FluentBitFilterFlowName],
-		DNS:  cm.Data[render.FluentBitFilterDNSName],
+	return &rlogcollector.FluentBitFilters{
+		Flow: cm.Data[rlogcollector.FluentBitFilterFlowName],
+		DNS:  cm.Data[rlogcollector.FluentBitFilterDNSName],
 	}, nil
 }
 
-func getEksCloudwatchLogConfig(client client.Client, interval int32, region, group, prefix string) (*render.EksCloudwatchLogConfig, error) {
+func getEksCloudwatchLogConfig(client client.Client, interval int32, region, group, prefix string) (*rlogcollector.EksCloudwatchLogConfig, error) {
 	if region == "" {
 		return nil, fmt.Errorf("missing AWS region info")
 	}
@@ -837,24 +838,24 @@ func getEksCloudwatchLogConfig(client client.Client, interval int32, region, gro
 
 	secret := &corev1.Secret{}
 	secretNamespacedName := types.NamespacedName{
-		Name:      render.EksLogForwarderSecret,
+		Name:      rlogcollector.EksLogForwarderSecret,
 		Namespace: common.OperatorNamespace(),
 	}
 	if err := client.Get(context.Background(), secretNamespacedName, secret); err != nil {
 		if errors.IsNotFound(err) {
 			return nil, nil
 		}
-		return nil, fmt.Errorf("failed to read Secret %q: %s", render.EksLogForwarderSecret, err)
+		return nil, fmt.Errorf("failed to read Secret %q: %s", rlogcollector.EksLogForwarderSecret, err)
 	}
 
-	if len(secret.Data[render.EksLogForwarderAwsId]) == 0 ||
-		len(secret.Data[render.EksLogForwarderAwsKey]) == 0 {
+	if len(secret.Data[rlogcollector.EksLogForwarderAwsId]) == 0 ||
+		len(secret.Data[rlogcollector.EksLogForwarderAwsKey]) == 0 {
 		return nil, fmt.Errorf("incomplete Cloudwatch credentials")
 	}
 
-	return &render.EksCloudwatchLogConfig{
-		AwsId:         secret.Data[render.EksLogForwarderAwsId],
-		AwsKey:        secret.Data[render.EksLogForwarderAwsKey],
+	return &rlogcollector.EksCloudwatchLogConfig{
+		AwsId:         secret.Data[rlogcollector.EksLogForwarderAwsId],
+		AwsKey:        secret.Data[rlogcollector.EksLogForwarderAwsKey],
 		AwsRegion:     region,
 		GroupName:     group,
 		StreamPrefix:  prefix,
@@ -865,21 +866,21 @@ func getEksCloudwatchLogConfig(client client.Client, interval int32, region, gro
 func getSysLogCertificate(client client.Client) (certificatemanagement.CertificateInterface, error) {
 	cm := &corev1.ConfigMap{}
 	cmNamespacedName := types.NamespacedName{
-		Name:      render.SyslogCAConfigMapName,
+		Name:      rlogcollector.SyslogCAConfigMapName,
 		Namespace: common.OperatorNamespace(),
 	}
 	if err := client.Get(context.Background(), cmNamespacedName, cm); err != nil {
 		if errors.IsNotFound(err) {
-			log.Info(fmt.Sprintf("ConfigMap %q is not found, assuming syslog's certificate is signed by publicly trusted CA", render.SyslogCAConfigMapName))
+			log.Info(fmt.Sprintf("ConfigMap %q is not found, assuming syslog's certificate is signed by publicly trusted CA", rlogcollector.SyslogCAConfigMapName))
 			return nil, nil
 		}
-		return nil, fmt.Errorf("failed to read ConfigMap %q: %s", render.SyslogCAConfigMapName, err)
+		return nil, fmt.Errorf("failed to read ConfigMap %q: %s", rlogcollector.SyslogCAConfigMapName, err)
 	}
 	if len(cm.Data[corev1.TLSCertKey]) == 0 {
-		log.Info(fmt.Sprintf("ConfigMap %q does not have a field named %q, assuming syslog's certificate is signed by publicly trusted CA", render.SyslogCAConfigMapName, corev1.TLSCertKey))
+		log.Info(fmt.Sprintf("ConfigMap %q does not have a field named %q, assuming syslog's certificate is signed by publicly trusted CA", rlogcollector.SyslogCAConfigMapName, corev1.TLSCertKey))
 		return nil, nil
 	}
-	syslogCert := certificatemanagement.NewCertificate(render.SyslogCAConfigMapName, common.OperatorNamespace(), []byte(cm.Data[corev1.TLSCertKey]), nil)
+	syslogCert := certificatemanagement.NewCertificate(rlogcollector.SyslogCAConfigMapName, common.OperatorNamespace(), []byte(cm.Data[corev1.TLSCertKey]), nil)
 
 	return syslogCert, nil
 }
