+++ pkg/render/logcollector/daemonset.go
@@ -1,0 +1,386 @@
+// Copyright (c) 2019-2026 Tigera, Inc. All rights reserved.
+
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package logcollector
+
+import (
+	"fmt"
+	"strings"
+
+	appsv1 "k8s.io/api/apps/v1"
+	corev1 "k8s.io/api/core/v1"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"k8s.io/apimachinery/pkg/util/intstr"
+
+	operatorv1 "github.com/tigera/operator/api/v1"
+	"github.com/tigera/operator/pkg/render"
+	rcomponents "github.com/tigera/operator/pkg/render/common/components"
+	rmeta "github.com/tigera/operator/pkg/render/common/meta"
+	"github.com/tigera/operator/pkg/render/common/secret"
+	"github.com/tigera/operator/pkg/render/common/securitycontext"
+	"github.com/tigera/operator/pkg/tls/certificatemanagement"
+)
+
+func (c *fluentBitComponent) healthProbeHandler() corev1.ProbeHandler {
+	return corev1.ProbeHandler{
+		HTTPGet: &corev1.HTTPGetAction{
+			Path: "/api/v1/health",
+			Port: intstr.FromInt(FluentBitMetricsPort),
+		},
+	}
+}
+
+func (c *fluentBitComponent) securityContext(privileged bool) *corev1.SecurityContext {
+	if c.cfg.OSType == rmeta.OSTypeWindows {
+		return nil
+	}
+	return securitycontext.NewRootContext(privileged)
+}
+
+func (c *fluentBitComponent) daemonset() *appsv1.DaemonSet {
+	var terminationGracePeriod int64 = 0
+	// The rationale for this setting is that while there is no need for fluent-bit to be available, we want to avoid
+	// potentially negative consequences of an immediate roll-out on huge clusters.
+	maxUnavailable := intstr.FromInt(10)
+
+	annots := c.cfg.TrustedBundle.HashAnnotations()
+
+	if c.cfg.FluentBitKeyPair != nil {
+		annots[c.cfg.FluentBitKeyPair.HashAnnotationKey()] = c.cfg.FluentBitKeyPair.HashAnnotationValue()
+	}
+	if c.cfg.S3Credential != nil {
+		annots[s3CredentialHashAnnotation] = rmeta.AnnotationHash(c.cfg.S3Credential)
+	}
+	if c.cfg.SplkCredential != nil {
+		annots[splunkCredentialHashAnnotation] = rmeta.AnnotationHash(c.cfg.SplkCredential)
+	}
+	// Most LogCollector spec changes only alter the rendered ConfigMap (and the
+	// config is subPath-mounted, which kubelet never live-updates), so hash the
+	// rendered config into the pod template to force a rollout on change. This
+	// also covers user filters, which are inlined into the config.
+	annots[configHashAnnotation] = rmeta.AnnotationHash(c.renderFluentBitConf())
+	var initContainers []corev1.Container
+	if c.cfg.OSType == rmeta.OSTypeLinux {
+		initContainers = append(initContainers, corev1.Container{
+			Name:    "pos-migrator",
+			Image:   c.image,
+			Command: []string{"/usr/bin/pos-migrator"},
+			Env: []corev1.EnvVar{
+				{Name: "LOG_DIRS", Value: c.logDirsCSV()},
+			},
+			SecurityContext: c.securityContext(false),
+			VolumeMounts: []corev1.VolumeMount{
+				{MountPath: "/var/log/calico", Name: "var-log-calico"},
+			},
+		})
+	} else {
+		// Windows fluentd also kept tail positions (.pos files under the same
+		// mounted log dir), so the cutover migration applies there too. The
+		// Windows image ships the cross-compiled migrator; env overrides point
+		// it at the c:-prefixed mounts.
+		initContainers = append(initContainers, corev1.Container{
+			Name:    "pos-migrator",
+			Image:   c.image,
+			Command: []string{c.path("/fluent-bit/pos-migrator.exe")},
+			Env: []corev1.EnvVar{
+				{Name: "POS_DIR", Value: c.path("/var/log/calico")},
+				{Name: "DB_DIR", Value: c.path("/var/log/calico/calico-fluent-bit")},
+				{Name: "LOG_DIRS", Value: c.logDirsCSV()},
+			},
+			SecurityContext: c.securityContext(false),
+			VolumeMounts: []corev1.VolumeMount{
+				{MountPath: c.path("/var/log/calico"), Name: "var-log-calico"},
+			},
+		})
+	}
+	if c.cfg.FluentBitKeyPair != nil && c.cfg.FluentBitKeyPair.UseCertificateManagement() {
+		initContainers = append(initContainers, c.cfg.FluentBitKeyPair.InitContainer(LogCollectorNamespace, c.container().SecurityContext))
+	}
+
+	podTemplate := &corev1.PodTemplateSpec{
+		ObjectMeta: metav1.ObjectMeta{
+			Annotations: annots,
+		},
+		Spec: corev1.PodSpec{
+			NodeSelector:                  map[string]string{},
+			Tolerations:                   rmeta.TolerateAll,
+			ImagePullSecrets:              secret.GetReferenceList(c.cfg.PullSecrets),
+			TerminationGracePeriodSeconds: &terminationGracePeriod,
+			InitContainers:                initContainers,
+			Containers:                    []corev1.Container{c.container()},
+			Volumes:                       c.volumes(),
+			ServiceAccountName:            c.fluentBitNodeName(),
+		},
+	}
+
+	ds := &appsv1.DaemonSet{
+		TypeMeta: metav1.TypeMeta{Kind: "DaemonSet", APIVersion: "apps/v1"},
+		ObjectMeta: metav1.ObjectMeta{
+			Name:      c.fluentBitNodeName(),
+			Namespace: LogCollectorNamespace,
+		},
+		Spec: appsv1.DaemonSetSpec{
+			Template: *podTemplate,
+			UpdateStrategy: appsv1.DaemonSetUpdateStrategy{
+				RollingUpdate: &appsv1.RollingUpdateDaemonSet{
+					MaxUnavailable: &maxUnavailable,
+				},
+			},
+		},
+	}
+	if c.cfg.LogCollector != nil {
+		overrides := c.cfg.LogCollector.Spec.CalicoFluentBitDaemonSet
+		if overrides == nil {
+			// Deprecated alias: fluentdDaemonSet is honored for one release when
+			// the new field is unset, per the API contract. Container entries
+			// stored under the legacy "fluentd" name are translated to the
+			// renamed container so resource overrides keep applying.
+			overrides = translateLegacyFluentdOverrides(c.cfg.LogCollector.Spec.FluentdDaemonSet) //nolint:staticcheck // deliberate use of the deprecated alias field
+		}
+		if overrides != nil {
+			rcomponents.ApplyDaemonSetOverrides(ds, overrides)
+		}
+	}
+	render.SetNodeCriticalPod(&(ds.Spec.Template))
+	return ds
+}
+
+// translateLegacyFluentdOverrides maps a deprecated fluentdDaemonSet override
+// onto the renamed calico-fluent-bit pod: container/init-container entries that
+// still carry fluentd-era names (stored before the CRD enum was updated) are
+// renamed so ApplyDaemonSetOverrides matches them.
+func translateLegacyFluentdOverrides(legacy *operatorv1.FluentBitDaemonSet) *operatorv1.FluentBitDaemonSet {
+	if legacy == nil {
+		return nil
+	}
+	translated := legacy.DeepCopy()
+	if translated.Spec == nil || translated.Spec.Template == nil || translated.Spec.Template.Spec == nil {
+		return translated
+	}
+	for i, container := range translated.Spec.Template.Spec.Containers {
+		if container.Name == "fluentd" {
+			translated.Spec.Template.Spec.Containers[i].Name = "calico-fluent-bit"
+		}
+	}
+	for i, container := range translated.Spec.Template.Spec.InitContainers {
+		translated.Spec.Template.Spec.InitContainers[i].Name = strings.Replace(container.Name, "tigera-fluentd-prometheus-tls", FluentBitTLSSecretName, 1)
+	}
+	return translated
+}
+
+func (c *fluentBitComponent) container() corev1.Container {
+	envs := c.envvars()
+	volumeMounts := []corev1.VolumeMount{
+		{MountPath: c.path("/var/log/calico"), Name: "var-log-calico"},
+		{MountPath: c.path("/etc/fluent-bit/certs"), Name: certificatemanagement.TrustedCertConfigMapName},
+	}
+	if c.cfg.OSType == rmeta.OSTypeWindows {
+		// Windows containers cannot mount a single file (no subPath file
+		// mounts), so mount the whole ConfigMap as a directory. The Windows
+		// image keeps its own files under C:\fluent-bit, so c:\etc\fluent-bit
+		// shadows nothing.
+		volumeMounts = append(volumeMounts,
+			corev1.VolumeMount{MountPath: c.path("/etc/fluent-bit/conf"), Name: "fluent-bit-conf", ReadOnly: true})
+	} else {
+		// Mount only the rendered config as a single file (SubPath) so it does
+		// not shadow the image's /etc/fluent-bit directory, which ships
+		// plugins.conf, record_transformer.lua and the in_eks plugin.
+		volumeMounts = append(volumeMounts,
+			corev1.VolumeMount{MountPath: "/etc/fluent-bit/fluent-bit.yaml", Name: "fluent-bit-conf", SubPath: "fluent-bit.yaml", ReadOnly: true})
+	}
+
+	volumeMounts = append(volumeMounts, c.cfg.TrustedBundle.VolumeMounts(c.SupportedOSType())...)
+
+	if c.cfg.FluentBitKeyPair != nil {
+		volumeMounts = append(volumeMounts, c.cfg.FluentBitKeyPair.VolumeMount(c.SupportedOSType()))
+	}
+
+	if c.cfg.ManagedCluster {
+		volumeMounts = append(volumeMounts,
+			corev1.VolumeMount{
+				Name:      LinseedTokenVolumeName,
+				MountPath: c.path(LinseedVolumeMountPath),
+			})
+	}
+
+	return corev1.Container{
+		Name:    "calico-fluent-bit",
+		Image:   c.image,
+		Command: []string{c.binPath()},
+		Args:    []string{"-c", c.configPath()},
+		Env:     envs,
+		// On OpenShift Fluent Bit needs privileged access to access logs on host path volume
+		SecurityContext: c.securityContext(c.cfg.Installation.KubernetesProvider.IsOpenShift()),
+		VolumeMounts:    volumeMounts,
+		StartupProbe:    c.startup(),
+		LivenessProbe:   c.liveness(),
+		ReadinessProbe:  c.readiness(),
+		Ports: []corev1.ContainerPort{{
+			Name:          "metrics-port",
+			ContainerPort: FluentBitMetricsPort,
+		}},
+	}
+}
+
+func (c *fluentBitComponent) metricsService() *corev1.Service {
+	return &corev1.Service{
+		TypeMeta: metav1.TypeMeta{Kind: "Service", APIVersion: "v1"},
+		ObjectMeta: metav1.ObjectMeta{
+			Name:      c.fluentBitMetricsServiceName(),
+			Namespace: LogCollectorNamespace,
+			Labels:    map[string]string{"k8s-app": c.fluentBitNodeName()},
+		},
+		Spec: corev1.ServiceSpec{
+			Selector: map[string]string{"k8s-app": c.fluentBitNodeName()},
+			// Important: "None" tells Kubernetes that we want a headless service with
+			// no kube-proxy load balancer.  If we omit this then kube-proxy will render
+			// a huge set of iptables rules for this service since there's an instance
+			// on every node.
+			ClusterIP: "None",
+			Ports: []corev1.ServicePort{
+				{
+					Name:       FluentBitMetricsPortName,
+					Port:       int32(FluentBitMetricsPort),
+					TargetPort: intstr.FromInt(FluentBitMetricsPort),
+					Protocol:   corev1.ProtocolTCP,
+				},
+			},
+		},
+	}
+}
+
+func (c *fluentBitComponent) envvars() []corev1.EnvVar {
+	envs := []corev1.EnvVar{
+		{Name: "NODENAME", ValueFrom: &corev1.EnvVarSource{FieldRef: &corev1.ObjectFieldSelector{FieldPath: "spec.nodeName"}}},
+	}
+	// Additional stores are Linux-only (the Windows pipeline ships to Linseed
+	// only, matching the fluentd Windows variant), so their credentials are too.
+	if c.cfg.LogCollector.Spec.AdditionalStores != nil && c.cfg.OSType == rmeta.OSTypeLinux {
+		if s3 := c.cfg.LogCollector.Spec.AdditionalStores.S3; s3 != nil {
+			// The standard AWS credential env vars, which fluent-bit's native s3
+			// output reads via the AWS credential chain (the legacy AWS_KEY_ID /
+			// AWS_SECRET_KEY names were fluentd-config-only and are read by
+			// nothing in fluent-bit).
+			envs = append(envs,
+				corev1.EnvVar{
+					Name: "AWS_ACCESS_KEY_ID",
+					ValueFrom: &corev1.EnvVarSource{
+						SecretKeyRef: &corev1.SecretKeySelector{
+							LocalObjectReference: corev1.LocalObjectReference{Name: S3FluentBitSecretName},
+							Key:                  S3KeyIdName,
+						},
+					},
+				},
+				corev1.EnvVar{
+					Name: "AWS_SECRET_ACCESS_KEY",
+					ValueFrom: &corev1.EnvVarSource{
+						SecretKeyRef: &corev1.SecretKeySelector{
+							LocalObjectReference: corev1.LocalObjectReference{Name: S3FluentBitSecretName},
+							Key:                  S3KeySecretName,
+						},
+					},
+				},
+			)
+		}
+		if splunk := c.cfg.LogCollector.Spec.AdditionalStores.Splunk; splunk != nil {
+			envs = append(envs,
+				corev1.EnvVar{
+					Name: "SPLUNK_HEC_TOKEN",
+					ValueFrom: &corev1.EnvVarSource{
+						SecretKeyRef: &corev1.SecretKeySelector{
+							LocalObjectReference: corev1.LocalObjectReference{Name: SplunkFluentBitTokenSecretName},
+							Key:                  SplunkFluentBitSecretTokenKey,
+						},
+					},
+				},
+			)
+		}
+	}
+	return envs
+}
+
+// The startup probe uses the same action as the liveness probe, but with
+// a higher failure threshold and double the timeout to account for slow
+// networks.
+func (c *fluentBitComponent) startup() *corev1.Probe {
+	return &corev1.Probe{
+		ProbeHandler:     c.healthProbeHandler(),
+		TimeoutSeconds:   c.probeTimeout,
+		PeriodSeconds:    c.probePeriod,
+		FailureThreshold: 10,
+	}
+}
+
+func (c *fluentBitComponent) liveness() *corev1.Probe {
+	return &corev1.Probe{
+		ProbeHandler:   c.healthProbeHandler(),
+		TimeoutSeconds: c.probeTimeout,
+		PeriodSeconds:  c.probePeriod,
+	}
+}
+
+func (c *fluentBitComponent) readiness() *corev1.Probe {
+	return &corev1.Probe{
+		ProbeHandler:   c.healthProbeHandler(),
+		TimeoutSeconds: c.probeTimeout,
+		PeriodSeconds:  c.probePeriod,
+	}
+}
+
+func (c *fluentBitComponent) volumes() []corev1.Volume {
+	dirOrCreate := corev1.HostPathDirectoryOrCreate
+
+	volumes := []corev1.Volume{
+		{
+			Name: "var-log-calico",
+			VolumeSource: corev1.VolumeSource{
+				HostPath: &corev1.HostPathVolumeSource{
+					Path: c.volumeHostPath(),
+					Type: &dirOrCreate,
+				},
+			},
+		},
+		{
+			Name: "fluent-bit-conf",
+			VolumeSource: corev1.VolumeSource{
+				ConfigMap: &corev1.ConfigMapVolumeSource{
+					LocalObjectReference: corev1.LocalObjectReference{
+						Name: c.fluentBitConfConfigMapName(),
+					},
+				},
+			},
+		},
+	}
+	if c.cfg.FluentBitKeyPair != nil {
+		volumes = append(volumes, c.cfg.FluentBitKeyPair.Volume())
+	}
+	if c.cfg.ManagedCluster {
+		volumes = append(volumes,
+			corev1.Volume{
+				Name: LinseedTokenVolumeName,
+				VolumeSource: corev1.VolumeSource{
+					Secret: &corev1.SecretVolumeSource{
+						// Per-OS: the token controller provisions a secret per
+						// ServiceAccount (calico-fluent-bit /
+						// calico-fluent-bit-windows).
+						SecretName: fmt.Sprintf(LinseedTokenSecret, c.fluentBitNodeName()),
+						Items:      []corev1.KeyToPath{{Key: LinseedTokenKey, Path: LinseedTokenSubPath}},
+					},
+				},
+			})
+	}
+	volumes = append(volumes, render.TrustedBundleVolume(c.cfg.TrustedBundle))
+
+	return volumes
+}
