+++ pkg/render/windows.go
@@ -887,7 +887,7 @@ func (c *windowsComponent) windowsDaemonset(cniCfgMap *corev1.ConfigMap) *appsv1
 		ds.Spec.Template.Spec.InitContainers = append(ds.Spec.Template.Spec.InitContainers, c.cniContainer())
 	}
 
-	setNodeCriticalPod(&(ds.Spec.Template))
+	SetNodeCriticalPod(&(ds.Spec.Template))
 
 	if overrides := c.cfg.Installation.CalicoNodeWindowsDaemonSet; overrides != nil {
 		rcomp.ApplyDaemonSetOverrides(&ds, overrides)
